refactor: ♻️ Move general modules to root
