+++ src/legacy_puyo_tools/_math.py
